# domains/core-platform-services/auth-service/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered."
          },
          "400": {
            "description": "Validation error or user already exists."
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in."
          },
          "401": {
            "description": "Invalid credentials."
          }
        },
        "summary": "Login a user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login-google": {
      "post": {
        "operationId": "AuthController_loginWithGoogle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idToken": {
                    "type": "string"
                  }
                },
                "required": ["idToken"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in with Google."
          },
          "401": {
            "description": "Invalid or expired Google ID token."
          }
        },
        "summary": "Login a user via Google SSO",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/verify": {
      "post": {
        "operationId": "AuthController_verify",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session verified successfully."
          },
          "401": {
            "description": "Invalid authentication."
          }
        },
        "summary": "Verify user session via token or cookie",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Gateway",
    "description": "Gateway documentation for frontend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignUpDto": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "description": "First name of the user",
            "example": "John"
          },
          "lastname": {
            "type": "string",
            "description": "Last name of the user",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password",
            "example": "strongPassword123"
          },
          "globalRole": {
            "type": "string",
            "description": "Global role of the user",
            "enum": [
              "USER",
              "ADMIN"
            ],
            "example": "USER"
          },
          "profilePicture": {
            "type": "string",
            "description": "Profile picture URL",
            "example": "https://example.com/profile.jpg"
          },
          "bio": {
            "type": "string",
            "description": "Short bio of the user",
            "example": "I am a new user"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password",
            "example": "strongPassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}